Make d3 get data through json request.
Before all the data was embedded in the html, but this caused a long load time. Now there is an AJAX call that gets the data, which lets the html load faster (and not look broken).

There are some functions here that support per section json requests for grade distribution, but that was abandoned in favor of requesting the data for all the sections for a particular type of graph. Wanted to keep the code around for a bit at least though.