Adding styles to fix issue BLD-221.
If you make the window smaller, and keep reducing it's size, at some
point the controls will not fit, and some will be bumped to the next row.
When the widnow is resized below the width 1024, extra styling rules
will kick in to fix this issue. The word "Speed" from the speeds control
will be temporarirly hidden.